Share render code between Wayland and Wayland direct widgets
By creating a _GlRendererMixin class in gl_helpers.
