+++ picamera2/previews/gl_helpers.py
@@ -125,6 +125,82 @@ def build_camera_programs(transform):
     return program_image, program_overlay, vert_positions, overlay_texture
 
 
+class _GlRendererMixin:
+    """Shared rendering logic for Qt-managed GL contexts (QOpenGLWidget / QOpenGLWindow).
+
+    Provides _build_programs, _repaint, recalculate_viewport, and set_overlay.
+    The host class __init__ must set: picamera2, transform, bg_colour, keep_ar,
+    lock, buffers, stop_count, egl_display, max_texture_size, program_image,
+    program_overlay, overlay_texture, overlay_present, overlay_array, _overlay_dirty.
+    """
+
+    def _build_programs(self):
+        (self.program_image, self.program_overlay,
+         self._vertPositions, self.overlay_texture) = build_camera_programs(self.transform)
+
+    def _repaint(self, completed_request):
+        if completed_request and completed_request.request not in self.buffers:
+            if self.stop_count != self.picamera2.stop_count:
+                for _, buffer in self.buffers.items():
+                    glDeleteTextures(1, [buffer.texture])
+                self.buffers = {}
+                self.stop_count = self.picamera2.stop_count
+            self.buffers[completed_request.request] = Buffer(
+                self.egl_display, completed_request, self.max_texture_size)
+
+        if self._overlay_dirty and self.overlay_array is not None:
+            glBindTexture(GL_TEXTURE_2D, self.overlay_texture)
+            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR)
+            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR)
+            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE)
+            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE)
+            height, width, _ = self.overlay_array.shape
+            glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, width, height, 0,
+                         GL_RGBA, GL_UNSIGNED_BYTE, self.overlay_array)
+            self._overlay_dirty = False
+
+        x_off, y_off, w, h = self.recalculate_viewport()
+        glViewport(x_off, y_off, w, h)
+        glClearColor(*self.bg_colour)
+        glClear(GL_COLOR_BUFFER_BIT)
+
+        if completed_request:
+            buffer = self.buffers[completed_request.request]
+            glUseProgram(self.program_image)
+            glBindTexture(GL_TEXTURE_EXTERNAL_OES, buffer.texture)
+            glDrawArrays(GL_TRIANGLE_FAN, 0, 4)
+
+        if self.overlay_present:
+            glUseProgram(self.program_overlay)
+            glBindTexture(GL_TEXTURE_2D, self.overlay_texture)
+            glDrawArrays(GL_TRIANGLE_FAN, 0, 4)
+
+    def recalculate_viewport(self):
+        dpr = self.devicePixelRatioF()
+        window_w = int(self.width() * dpr)
+        window_h = int(self.height() * dpr)
+        stream_map = self.picamera2.stream_map
+        camera_config = self.picamera2.camera_config
+        if not self.keep_ar or not camera_config or camera_config['display'] is None:
+            return 0, 0, window_w, window_h
+        image_w = stream_map[camera_config['display']].configuration.size.width
+        image_h = stream_map[camera_config['display']].configuration.size.height
+        if image_w * window_h > window_w * image_h:
+            w = window_w
+            h = w * image_h // image_w
+        else:
+            h = window_h
+            w = h * image_w // image_h
+        return (window_w - w) // 2, (window_h - h) // 2, w, h
+
+    def set_overlay(self, overlay):
+        with self.lock:
+            self.overlay_array = overlay
+            self.overlay_present = overlay is not None
+            self._overlay_dirty = overlay is not None
+        self.update()
+
+
 def str_to_fourcc(str):
     assert len(str) == 4
     fourcc = 0
